requirements.txt
dlib
pycryptodomex
pyside6
cv2
